+++ src/error.rs
@@ -1,0 +1,25 @@
+use crate::Span;
+
+/// Error returned in case of invalid beancount syntax found
+///
+/// # Example
+/// ```
+/// let result = beancount_parser_2::parse::<f64>("2022-05-21 oops");
+/// assert!(result.is_err());
+/// let error = result.unwrap_err();
+/// assert_eq!(error.line_number(), 1);
+/// ```
+#[derive(Debug)]
+pub struct Error<'a>(Span<'a>);
+
+impl<'a> Error<'a> {
+    pub(crate) fn new(span: Span<'a>) -> Self {
+        Self(span)
+    }
+
+    /// Line number at which the error was found in the input
+    #[must_use]
+    pub fn line_number(&self) -> u32 {
+        self.0.location_line()
+    }
+}